backend/app/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
psycopg2-binary
sqlalchemy
passlib[bcrypt]
bcrypt
python-jose[cryptography]
email-validator
watchfiles
python-multipart